исправлено наименование пакетов на ru.praktikum.services.qa.scooter
