+++ src/main/java/ru/praktikum/services/qa/scooter/constants/Color.java
@@ -1,4 +1,4 @@
-package ru.praktikumServices.QaScooter.constants;
+package ru.praktikum.services.qa.scooter.constants;
 
 public class Color {
     public static final String COLOR_GREY = "grey";
