+++ src/main/java/ru/praktikum/services/qa/scooter/constants/Result.java
@@ -1,4 +1,4 @@
-package ru.praktikumServices.QaScooter.constants;
+package ru.praktikum.services.qa.scooter.constants;
 
 public class Result {
     public static final String RESULT_ORDER_SCOOTER_POSITIVE = "Заказ оформлен";
