+++ src/main/java/ru/praktikum/services/qa/scooter/pom/OrderPage.java
@@ -1,4 +1,4 @@
-package ru.praktikumServices.QaScooter.pom;
+package ru.praktikum.services.qa.scooter.pom;
 
 import org.openqa.selenium.*;
